requirements.txt
PyQt5
pillow
mammoth
pyinstaller